refactor: extract file upload from swap component
